Bug fixing
The new implementation of the insertion in the hash table was provoking an error in the freeing of the table. So i went back to the previous one to assure the correct execution of the code.

Also a new function was added (fmod) and now you can use `^` to call the pow function, and `%` tocall the mod function.